feat: close sidebar on mobile
