+++ src/app/dashboard/components/dashboard-layout.tsx
@@ -14,16 +14,7 @@ import AuthComponent from './auth-component/auth-component';
 import MenuContainer from './sidebar/menu-container';
 import Sidebar from './sidebar/sidebar';
 
-type Props = {
-  menuItems: ReactNode;
-  secondMenuItems?: ReactNode;
-};
-
-export default function DashboardLayout({
-  children,
-  menuItems,
-  secondMenuItems,
-}: PropsWithChildren<Props>) {
+export default function DashboardLayout({ children }: PropsWithChildren) {
   return (
     <Stack
       direction={{
@@ -33,46 +24,7 @@ export default function DashboardLayout({
       alignItems="stretch"
       sx={{ minHeight: '100%' }}
     >
-      <Sidebar>
-        <Box
-          sx={{
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-          }}
-        >
-          <Box sx={{ pl: 0.5 }}>
-            <Logo href={routes.dashboard.storage} />
-          </Box>
-          {isSandbox && (
-            <Tooltip
-              title="You are on the Gateway Sandbox. The data is temporary and will expire in 60 days."
-              placement="right"
-              arrow
-            >
-              <Chip
-                color="warning"
-                size="small"
-                variant="filled"
-                label="Sandbox"
-              />
-            </Tooltip>
-          )}
-        </Box>
-        <Stack
-          justifyContent="space-between"
-          flexDirection="column"
-          sx={{
-            height: 'calc(100% - 40px)',
-          }}
-        >
-          <MenuContainer
-            menuItems={menuItems}
-            secondMenuItems={secondMenuItems}
-          />
-          <AuthComponent id="profile-button" controlId="profile-menu" />
-        </Stack>
-      </Sidebar>
+      <Sidebar />
       <Box
         width="100%"
         sx={{
