+++ src/app/dashboard/components/sidebar/dashboard-menu-items.tsx
@@ -2,7 +2,7 @@ import DataModelOutlinedIcon from '@/components/icons/data-model-outlined';
 import StorageIcon from '@/components/icons/storage';
 import routes from '@/constants/routes';
 
-import { GTWMenuItemSettings } from './menu-item';
+import { GTWMenuItemSettings } from '../menu-item';
 
 /**
  * List all menu items of the user dashboard
