+++ src/app/dashboard/layout.tsx
@@ -1,9 +1,7 @@
 import { Metadata } from 'next';
 import { PropsWithChildren } from 'react';
 
-import DashboardUserDeveloperMenuListItems from './components/dashboard-developer-menu-list-items';
 import DashboardLayout from './components/dashboard-layout';
-import DashboardUserMenuListItems from './components/dashboard-menu-list-items';
 import DashboardErrorBoundary from './components/error';
 
 export const metadata: Metadata = {
@@ -16,12 +14,7 @@ export const metadata: Metadata = {
 export default function DashboardUserLayout({ children }: PropsWithChildren) {
   return (
     <DashboardErrorBoundary>
-      <DashboardLayout
-        menuItems={<DashboardUserMenuListItems />}
-        secondMenuItems={<DashboardUserDeveloperMenuListItems />}
-      >
-        {children}
-      </DashboardLayout>
+      <DashboardLayout>{children}</DashboardLayout>
     </DashboardErrorBoundary>
   );
 }
